change `CalculatePipelineResources` into resource provider
