+++ internal/tools/pipeline/services/pipelinesvc/create.go
@@ -369,7 +369,7 @@ func (s *PipelineSvc) CreatePipelineGraph(p *spec.Pipeline) (newStages []spec.Pi
 		}
 
 		// calculate pipeline applied resource after all snippetTask created
-		pipelineAppliedResources, err := s.calculatePipelineResources(pipelineYml, p)
+		pipelineAppliedResources, err := s.resource.CalculatePipelineResources(pipelineYml, p)
 		if err != nil {
 			return nil, apierrors.ErrCreatePipelineGraph.InternalError(
 				fmt.Errorf("failed to search pipeline action resrouces, err: %v", err))
