+++ internal/tools/pipeline/services/pipelinesvc/create_task_test.go
@@ -16,7 +16,6 @@ package pipelinesvc
 
 import (
 	"fmt"
-	"reflect"
 	"testing"
 	"time"
 
@@ -26,205 +25,8 @@ import (
 	"github.com/erda-project/erda/bundle"
 	"github.com/erda-project/erda/internal/tools/pipeline/spec"
 	"github.com/erda-project/erda/pkg/http/httpclient"
-	"github.com/erda-project/erda/pkg/parser/diceyml"
 	"github.com/erda-project/erda/pkg/parser/pipelineyml"
 )
-
-func genActionWithRes(cpu, maxCPU float64, memoryMB int) *pipelineyml.Action {
-	return &pipelineyml.Action{
-		Resources: pipelineyml.Resources{
-			CPU:    cpu,
-			MaxCPU: maxCPU,
-			Mem:    memoryMB,
-		},
-	}
-}
-
-func genActionDefineWithRes(cpu, maxCPU float64, memoryMB, maxMemoryMB int) *diceyml.Job {
-	return &diceyml.Job{
-		Resources: diceyml.Resources{
-			CPU:    cpu,
-			MaxCPU: maxCPU,
-			Mem:    memoryMB,
-			MaxMem: maxMemoryMB,
-		},
-	}
-}
-
-func Test_calculateNormalTaskLimitResource(t *testing.T) {
-	type args struct {
-		action       *pipelineyml.Action
-		actionDefine *diceyml.Job
-		defaultRes   apistructs.PipelineAppliedResource
-	}
-	tests := []struct {
-		name string
-		args args
-		want apistructs.PipelineAppliedResource
-	}{
-		{
-			name: "all defined",
-			args: args{
-				action:       genActionWithRes(1, 0, 2049),
-				actionDefine: genActionDefineWithRes(2, 3, 1024, 2048),
-				defaultRes:   apistructs.PipelineAppliedResource{CPU: 0.1, MemoryMB: 32},
-			},
-			want: apistructs.PipelineAppliedResource{
-				CPU:      3,
-				MemoryMB: 2049,
-			},
-		},
-		{
-			name: "no custom action resources defined",
-			args: args{
-				action:       genActionWithRes(0, 0, 0),
-				actionDefine: genActionDefineWithRes(2, 3, 1024, 2048),
-				defaultRes:   apistructs.PipelineAppliedResource{CPU: 0.1, MemoryMB: 32},
-			},
-			want: apistructs.PipelineAppliedResource{
-				CPU:      3,
-				MemoryMB: 2048,
-			},
-		},
-		{
-			name: "all undefined",
-			args: args{
-				action:       genActionWithRes(0, 0, 0),
-				actionDefine: genActionDefineWithRes(0, 0, 0, 0),
-				defaultRes:   apistructs.PipelineAppliedResource{CPU: 0.1, MemoryMB: 32},
-			},
-			want: apistructs.PipelineAppliedResource{
-				CPU:      0.1,
-				MemoryMB: 32,
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got := calculateNormalTaskLimitResource(tt.args.action, tt.args.actionDefine, tt.args.defaultRes); !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("calculateNormalTaskLimitResource() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-func Test_calculateOversoldTaskLimitResource(t *testing.T) {
-	type args struct {
-		action       *pipelineyml.Action
-		actionDefine *diceyml.Job
-		defaultRes   apistructs.PipelineAppliedResource
-	}
-	tests := []struct {
-		name string
-		args args
-		want apistructs.PipelineAppliedResource
-	}{
-		{
-			name: "cpu lower than define and default",
-			args: args{
-				action:       genActionWithRes(0.1, 0, 2049),
-				actionDefine: genActionDefineWithRes(0.2, 0.4, 1024, 2048),
-				defaultRes:   apistructs.PipelineAppliedResource{CPU: 0.5, MemoryMB: 32},
-			},
-			want: apistructs.PipelineAppliedResource{
-				CPU:      0.8,
-				MemoryMB: 2049,
-			},
-		},
-		{
-			name: "cpu bigger than define but lower than default",
-			args: args{
-				action:       genActionWithRes(0.2, 0, 0),
-				actionDefine: genActionDefineWithRes(0.1, 0.5, 1024, 2048),
-				defaultRes:   apistructs.PipelineAppliedResource{CPU: 1, MemoryMB: 32},
-			},
-			want: apistructs.PipelineAppliedResource{
-				CPU:      1,
-				MemoryMB: 2048,
-			},
-		},
-		{
-			name: "cpu bigger than define and default",
-			args: args{
-				action:       genActionWithRes(3, 0, 0),
-				actionDefine: genActionDefineWithRes(2, 0, 0, 0),
-				defaultRes:   apistructs.PipelineAppliedResource{CPU: 1, MemoryMB: 32},
-			},
-			want: apistructs.PipelineAppliedResource{
-				CPU:      2,
-				MemoryMB: 32,
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got := calculateOversoldTaskLimitResource(calculateNormalTaskLimitResource(tt.args.action, tt.args.actionDefine, tt.args.defaultRes), apistructs.PipelineOverSoldResource{
-				CPURate: 2,
-				MaxCPU:  2,
-			}); !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("calculateNormalTaskLimitResource() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-func Test_calculateNormalTaskRequestResource(t *testing.T) {
-	type args struct {
-		action       *pipelineyml.Action
-		actionDefine *diceyml.Job
-		defaultRes   apistructs.PipelineAppliedResource
-	}
-
-	tests := []struct {
-		name string
-		args args
-		want apistructs.PipelineAppliedResource
-	}{
-		{
-			name: "all defined",
-			args: args{
-				action:       genActionWithRes(1, 0, 2049),
-				actionDefine: genActionDefineWithRes(2, 3, 1024, 2048),
-				defaultRes:   apistructs.PipelineAppliedResource{CPU: 0.1, MemoryMB: 32},
-			},
-			want: apistructs.PipelineAppliedResource{
-				CPU:      1,
-				MemoryMB: 2049,
-			},
-		},
-		{
-			name: "no custom action resources defined",
-			args: args{
-				action:       genActionWithRes(0, 0, 0),
-				actionDefine: genActionDefineWithRes(2, 3, 1024, 2048),
-				defaultRes:   apistructs.PipelineAppliedResource{CPU: 0.1, MemoryMB: 32},
-			},
-			want: apistructs.PipelineAppliedResource{
-				CPU:      2,
-				MemoryMB: 1024,
-			},
-		},
-		{
-			name: "all undefined",
-			args: args{
-				action:       genActionWithRes(0, 0, 0),
-				actionDefine: genActionDefineWithRes(0, 0, 0, 0),
-				defaultRes:   apistructs.PipelineAppliedResource{CPU: 0.1, MemoryMB: 32},
-			},
-			want: apistructs.PipelineAppliedResource{
-				CPU:      0.1,
-				MemoryMB: 32,
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got := calculateNormalTaskRequestResource(tt.args.action, tt.args.actionDefine, tt.args.defaultRes); !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("calculateNormalTaskRequestResource() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
 
 func TestPipelineSvc_judgeTaskExecutor(t *testing.T) {
 	type args struct {
